requirement.txt
# python_version >= "3.10.13"
python-dotenv==0.21.0
notion-client==2.2.1
pandas==2.1.4